Bug fix: Record number of eligible members at beginning of a vote resp. election
The old code calculated the set of eligible members just in time when
summarizing the election or vote results. This would have been inconsistent
with members becoming non-eligible because they expire and would
have made old results invalid.

This fixes it by saving the number of members eligible to vote just when
opening a proposal vote / membership election.