More dependencies as further preparation for object delete
Make proposal votes depend on their result (if it exists) and
make raw files depend on their metadata (if it exists).

Should allow to use (the inverse of) dependencies() to figure out
which objects are safe to delete without breaking the structure of
the voting DAG.

Note that this makes the .dependencies() graph contain circles now.
This is not a problem however, as the zip-functionality (as its only
user so far) was already dealing with that possibility.