Implementation of object deletion
A votemaster can now delete objects by using the "delete-objects" action.

This functionality allows to delete a set of objects which:

 - Is self contained in the sense that no other objects refer to the listed objects.
 - Is not the current member list.

This can be used for example to remove botched or stale proposal uploads.

It should be used with care, of course.

In case the object list is incompletely (there are objects depending
on deleted objects that are not listed), an error page listing all those
objects is rendered.

Also includes tests of this new feature.

Closes #7.