Support for multi-action
A multi-action is a list of actions with one single common signature that
can be submitted to the system all at once. It will be used for such things as
letting a member vote on all proposals or all new members at once.

Comes with test cases, and a new API endpoint /multi-action.

Multiple actions are to be separated by the string "\n@@@@@\n" and then signed.

A multi-action needs to consist of at least one regular action.

The good old regular "Action" type can now have an empty signature but
then has a pointer that will point to such a new "MultiAction" type as being
part of that.