Support for casting votes on multiple proposals or members all at once
Also adds the metadata (BUIP designation) to the vote action
string as a comment. Add action success template, test cases
and the necessary changes to common.html.

New form lists all open proposals to vote on.

Also add links to the main page for multi-voting on new members and
proposals.

Closes #7.